Use double precision translate.
It is unintuitive (since translate is in mesh units), but ncessary
for interlacing (A translation of 0.5).

Added 'shift' method to Affine. Not a good name, but the verb form
is already taken (translate)

+++ pmesh/_window_imp.h
@@ -25,7 +25,7 @@ struct FastPMPainter {
     int support;
     int ndim;
     double scale[32]; /* scale from position to grid units */
-    ptrdiff_t translate[32]; /* translate in grid units */
+    double translate[32]; /* translate in grid units */
     ptrdiff_t Nmesh[32]; /* periodicity */
 
     void * canvas;
